+++ src/test.ts
@@ -1,9 +1,0 @@
-import 'k8w-extend-native';
-import { cmdDoc } from "./commands/doc";
-
-cmdDoc({
-    input: './ptl',
-    output: 'docs',
-    ignore: undefined,
-    verbose: true
-})